feat: soft delete

+++ libs/nest-core/src/mongo/mongo.repository.ts
@@ -23,13 +23,15 @@ export abstract class MongoRepository<TEntity extends BaseSchema> {
 
   constructor(public readonly model: Model<TEntity>) {}
 
-  public async findById(id: string) {
-    const item: TEntity | null = await this.model.findById(id).lean()
+  public async findById(id: string, isDeleted: boolean | null = false) {
+    if (isDeleted === null) {
+      return this.model.findById(id).lean()
+    }
 
-    return item
+    return this.findOne({ filter: { _id: id }, isDeleted })
   }
 
-  public async findOne(options?: {
+  public async findOne(options: {
     filter: FilterQuery<TEntity>
     projection?:
       | keyof TEntity
@@ -37,12 +39,22 @@ export abstract class MongoRepository<TEntity extends BaseSchema> {
       | Record<keyof TEntity, number | boolean | object>
     populates?: Array<{
       path: keyof TEntity
+      isDeleted?: boolean | null
       fields?: Array<string>
+      match?: FilterQuery<TEntity> | (() => FilterQuery<TEntity>)
     }>
+    isDeleted?: boolean | null
   }) {
-    const { filter, projection, populates } = options ?? {}
+    const { filter, projection, populates, isDeleted } = options ?? {}
+    let query: ReturnType<typeof this.model.findOne>
 
-    const query = this.model.findOne(filter)
+    if (isDeleted === null) {
+      query = this.model.findOne(filter)
+    } else {
+      query = this.model.findOne({
+        $and: [filter, { isDeleted: isDeleted ?? false }],
+      })
+    }
 
     if (projection != undefined) {
       // @ts-ignore
@@ -57,8 +69,19 @@ export abstract class MongoRepository<TEntity extends BaseSchema> {
     return item
   }
 
-  public async exists(filter: FilterQuery<TEntity>) {
-    return !isNil((await this.model.exists(filter).lean())?._id)
+  public async exists(
+    filter: FilterQuery<TEntity>,
+    isDeleted: boolean | null = false,
+  ) {
+    let filterObj = filter
+
+    if (isDeleted !== null) {
+      filterObj = {
+        $and: [filter, { isDeleted }],
+      }
+    }
+
+    return !isNil((await this.model.exists(filterObj).lean())?._id)
   }
 
   public async create(data: Omit<TEntity, keyof BaseSchema>): Promise<TEntity> {
@@ -67,18 +90,33 @@ export abstract class MongoRepository<TEntity extends BaseSchema> {
     return item.toObject<TEntity>(this.toObjectOptions)
   }
 
-  public async count(filter: FilterQuery<TEntity>) {
-    return await this.model.countDocuments(filter).exec()
+  public async count(
+    filter: FilterQuery<TEntity>,
+    isDeleted: boolean | null = false,
+  ) {
+    let filterObj = filter
+
+    if (isDeleted !== null) {
+      filterObj = {
+        $and: [filter, { isDeleted }],
+      }
+    }
+
+    return await this.model.countDocuments(filterObj).exec()
   }
 
   private populate(
     query: any,
     populates: Array<{
       path: keyof TEntity
+      isDeleted?: boolean | null
       fields?: Array<string>
+      match?: FilterQuery<TEntity> | (() => FilterQuery<TEntity>)
     }>,
   ) {
     populates.forEach((populate) => {
+      const isDeleted =
+        populate.isDeleted !== null ? populate.isDeleted ?? false : null
       if (populate.path) {
         let populateObj = { path: populate.path }
 
@@ -86,12 +124,34 @@ export abstract class MongoRepository<TEntity extends BaseSchema> {
           populateObj['select'] = populate.fields.join(' ')
         }
 
+        if (populate.match !== undefined) {
+          let matchObject: typeof populate.match
+
+          if (typeof populate.match === 'function') {
+            matchObject = populate.match()
+          } else {
+            matchObject = populate.match
+          }
+
+          if (isDeleted !== null) {
+            matchObject = {
+              $and: [{ isDeleted }, matchObject],
+            }
+          }
+
+          populateObj['match'] = matchObject
+        } else {
+          if (isDeleted !== null) {
+            populateObj['match'] = { isDeleted }
+          }
+        }
+
         query = query.populate(populateObj)
       }
     })
   }
 
-  public async search(options?: {
+  public async search(options: {
     offset?: number
     limit?: number
     filter?: FilterQuery<TEntity>
@@ -102,12 +162,25 @@ export abstract class MongoRepository<TEntity extends BaseSchema> {
       | Record<keyof TEntity, number | boolean | object>
     populates?: Array<{
       path: keyof TEntity
+      isDeleted?: boolean | null
       fields?: Array<string>
+      match?: FilterQuery<TEntity> | (() => FilterQuery<TEntity>)
     }>
+    isDeleted?: boolean | null
   }) {
     const { offset, limit, filter, sort, projection, populates } = options ?? {}
+    const isDeleted =
+      options.isDeleted !== null ? options.isDeleted ?? false : null
 
-    const query = this.model.find(filter ?? {})
+    let filterObj = filter ?? {}
+
+    if (isDeleted !== null) {
+      filterObj = {
+        $and: [filterObj, { isDeleted }],
+      }
+    }
+
+    const query = this.model.find(filterObj)
 
     if (sort) {
       query.sort(sort)
@@ -139,7 +212,7 @@ export abstract class MongoRepository<TEntity extends BaseSchema> {
     }
   }
 
-  public async updateById(
+  public async updateByIdIgnoreSoftDelete(
     id: string,
     data: UpdateQuery<TEntity>,
     options?: QueryOptions<TEntity>,
@@ -155,9 +228,18 @@ export abstract class MongoRepository<TEntity extends BaseSchema> {
     filter: FilterQuery<TEntity>,
     data: UpdateQuery<TEntity>,
     options?: QueryOptions<TEntity>,
+    isDeleted: boolean | null = false,
   ) {
+    let filterObj = filter
+
+    if (isDeleted !== null) {
+      filterObj = {
+        $and: [filter, { isDeleted }],
+      }
+    }
+
     const item: TEntity | null = await this.model
-      .findOneAndUpdate(filter, data, options)
+      .findOneAndUpdate(filterObj, data, options)
       .lean()
 
     return item
@@ -167,21 +249,41 @@ export abstract class MongoRepository<TEntity extends BaseSchema> {
     filter: FilterQuery<TEntity>,
     data: UpdateQuery<TEntity>,
     options?: QueryOptions<TEntity>,
+    isDeleted: boolean | null = false,
   ) {
-    await this.model.updateMany(filter, data, options).lean()
+    let filterObj = filter
+
+    if (isDeleted !== null) {
+      filterObj = {
+        $and: [filter, { isDeleted }],
+      }
+    }
+
+    await this.model.updateMany(filterObj, data, options).lean()
   }
 
-  public async deleteById(id: string) {
-    const item: TEntity = await this.model.findByIdAndDelete(id).lean()
+  public async deleteById(id: string, softDelete = true) {
+    if (!softDelete) {
+      return this.model.findByIdAndDelete(id).lean<TEntity>()
+    }
 
-    return item
+    return this.model
+      .findByIdAndUpdate(id, { isDeleted: true, deletedAt: new Date() })
+      .lean<TEntity>()
   }
 
-  public async deleteMany(filter: FilterQuery<TEntity>) {
-    await this.model.deleteMany(filter).exec()
+  public async deleteMany(filter: FilterQuery<TEntity>, softDelete = true) {
+    if (!softDelete) {
+      this.model.deleteMany(filter).exec()
+      return
+    }
+
+    await this.model
+      .updateMany(filter, { isDeleted: true, deletedAt: new Date() })
+      .exec()
   }
 
-  public async bulkUpsert(
+  public async bulkUpsertIgnoreSoftDelete(
     docs: unknown[],
     upsert?: {
       conditions?: Array<string>
@@ -206,11 +308,11 @@ export abstract class MongoRepository<TEntity extends BaseSchema> {
     await this.model.bulkWrite(writes)
   }
 
-  public async bulkWrite(data: Array<object>) {
+  public async bulkWriteIgnoreSoftDelete(data: Array<object>) {
     await this.model.bulkWrite(data as any)
   }
 
-  public async aggregate<Result = any>(
+  public async aggregateIgnoreSoftDelete<Result = any>(
     pipelines: Array<PipelineStage>,
     allowDiskUse = false,
   ): Promise<Array<Result>> {
